fix: rename logging.py to avoid standard library conflict
A local module named `logging.py` in `src/soso/` is shadowing Python's
standard `logging` library.

This causes a `ModuleNotFoundError` when dependencies (like `requests`
-> `urllib3`) tries to import submodules from the standard library
(e.g., `logging.config`). Python is incorrectly looking for `config`
inside our local `logging.py` file instead of the built-in module.

Fix this by moving `src/soso/logging.py` functions to
`src/soso/utilities.py` and update all internal import statements to use
the new module.
